requirements.txt
click==7.1.2
dataclasses==0.8
Flask==1.1.2
future==0.18.2
glob2==0.7
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
Pillow==8.1.2
pkg-resources==0.0.0
resnet-pytorch==0.2.0
scipy==1.5.4
torch==1.7.0
torchvision==0.8.1
tqdm==4.59.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
